(#206) add testes do usuario
